Fix: update household queries to use 'simulation_instance_id' instead of 'simulation_instance'
