+++ food_access_model/preprocessing/get_data.py
@@ -8,6 +8,7 @@ from datetime import datetime
 from io import BytesIO
 from typing import List, Tuple, Optional, Callable, Any
 from zipfile import ZipFile
+import sys
 
 #  Third-Party Library Imports (External packages you install with pip):
 import numpy as np
@@ -45,8 +46,15 @@ from household_constants import (
 
 load_dotenv()
 
+# Configure logging
+logging.basicConfig(
+    level=logging.INFO,
+    format='%(asctime)s [%(levelname)s] %(message)s',
+    handlers=[logging.StreamHandler(sys.stdout)]
+)
+
 COUNTY_CODE = FIPSCODE[2:]                  
-STATE_CODE = FIPSCODE[:2]   
+STATE_CODE = FIPSCODE[:2]
 
 PASS = os.getenv("DB_PASS")
 if not PASS:
@@ -151,10 +159,6 @@ def fetch_county_data(
             county_data = pd.DataFrame(data[1:], columns= data[0])
     return county_data
 
-# Initialize county data using the function
-COUNTY_DATA = fetch_county_data(households_key_list, YEAR, STATE_CODE, COUNTY_CODE, APIKEY)
-
-
 def load_and_merge_geodata(
     year: str,
     state_code: str,
@@ -224,13 +228,6 @@ def load_and_merge_geodata(
 
     return merged_data
 
-# Run and create merged geospatial data
-data = load_and_merge_geodata(YEAR, STATE_CODE, COUNTY_CODE, COUNTY_DATA)
-
-# Create spatial index from geometry
-tract_index = STRtree(data["geometry"])
-
-
 def initialize_database_tables(
     host: str,
     database: str,
@@ -287,13 +284,18 @@ def initialize_database_tables(
     '''
     create_food_stores_query = '''
     CREATE TABLE food_stores (
+        simulation_instance UUID,
+        simulation_step INTEGER,
         shop VARCHAR(15),
         geometry TEXT,
-        name VARCHAR(50)
+        name VARCHAR(50),
+        store_id INTEGER
     );
     '''
     cursor.execute(create_roads_query)
     cursor.execute(create_food_stores_query)
+    
+    return connection, cursor
 
 
 def process_road_network(
@@ -327,25 +329,42 @@ def process_road_network(
     if "service" not in gdf_edges.columns:
         gdf_edges["service"] = None
     gdf_edges = gdf_edges[["name", "highway", "length", "geometry", "service"]]
+    
+    # Fetch water bodies to prevent house placement in rivers/lakes
+    try:
+        water_features = ox.features.features_from_point(
+            center_point,
+            dist=dist,
+            tags={"natural": ["water"], "waterway": ["river", "stream", "canal"]}
+        )
+        if not water_features.empty:
+            water_features = water_features.to_crs("epsg:3857")
+            for row in water_features.itertuples():
+                if hasattr(row, 'geometry') and row.geometry is not None:
+                    # Buffer water bodies slightly to create exclusion zones
+                    map_elements.append(row.geometry.buffer(10))
+            logging.info(f"Added {len(water_features)} water bodies to exclusion zones")
+    except Exception as e:
+        logging.warning(f"Could not fetch water features: {e}. Continuing without water filtering.")
 
 
-    for _, row in gdf_edges.itertuples():
-        if row["highway"] in ["residential", "living_street"]:
-            housing_areas.append(row["geometry"].buffer(30))
-            map_elements.append(row["geometry"].buffer(3))
-        elif row["service"] == "alley":
-            housing_areas.append(row["geometry"].buffer(30))
-            map_elements.append(row["geometry"].buffer(3))
-        elif row["highway"] == "motorway":
-            map_elements.append(row["geometry"].buffer(100))
-        elif row["highway"] == "trunk":
-            map_elements.append(row["geometry"].buffer(30))
-        elif row["highway"] == "primary":
-            map_elements.append(row["geometry"].buffer(10))
-        elif row["highway"] == "secondary":
-            map_elements.append(row["geometry"].buffer(10))
-        elif isinstance(row["geometry"], LineString):
-            map_elements.append(row["geometry"])
+    for row in gdf_edges.itertuples():
+        if row.highway in ["residential", "living_street"]:
+            housing_areas.append(row.geometry.buffer(30))
+            map_elements.append(row.geometry.buffer(3))
+        elif row.service == "alley":
+            housing_areas.append(row.geometry.buffer(30))
+            map_elements.append(row.geometry.buffer(3))
+        elif row.highway == "motorway":
+            map_elements.append(row.geometry.buffer(100))
+        elif row.highway == "trunk":
+            map_elements.append(row.geometry.buffer(30))
+        elif row.highway == "primary":
+            map_elements.append(row.geometry.buffer(10))
+        elif row.highway == "secondary":
+            map_elements.append(row.geometry.buffer(10))
+        elif isinstance(row.geometry, LineString):
+            map_elements.append(row.geometry)
 
     gdf_edges["length"] = gdf_edges["length"].astype(int)
     gdf_edges["geometry"] = gdf_edges["geometry"].astype(str)
@@ -354,20 +373,13 @@ def process_road_network(
 
     return map_elements, housing_areas, data_tuples
 
-
-# Run the road network processor
-map_elements, housing_areas, data_tuples = process_road_network(CENTER_POINT, DIST)
-
-# Prepare the SQL insertion query
-roads_query = "INSERT INTO roads (name, highway, length, geometry, service) VALUES %s"
-
-
+# remove cursor, no need now you aren't inserting data
+# return the list of store tuples additionally
 def process_food_stores(
     center_point: Tuple[float, float],
     dist: float,
     map_elements: List[geometry.base.BaseGeometry],
-    cursor: psycopg2.extensions.cursor
-) -> Tuple[STRtree, List] : 
+) -> Tuple[STRtree, List, List[Tuple]] : 
     """
     Retrieve and process food store locations from OpenStreetMap, convert them into geometric shapes,
     and insert them into the database.
@@ -375,7 +387,7 @@ def process_food_stores(
     Args:
         center_point (Tuple[float, float]): Latitude and longitude for the area of interest.
         dist (float): Distance in meters for the search radius (3x for food stores).
-        map_elements (List[BaseGeometry]): List to append buffered polygons representing stores.
+        map_elements (List[BaseGeometry]): List to append buffered polygons representing stores
         cursor (psycopg2.extensions.cursor): Cursor for executing database insert queries.
 
     Returns:
@@ -393,15 +405,14 @@ def process_food_stores(
     features = features.to_crs("epsg:3857")
     features = features[["shop", "geometry", "name"]]
 
-    store_tuples_strPoly: List[Tuple[str, str, str]] = []
-    store_tuples_Poly: List[Tuple[str, str, str]] = []
+    store_tuples_strPoly: List[Tuple] = []
+    store_tuples_Poly: List[Tuple] = []
 
-    food_stores_query = "INSERT INTO food_stores (shop, geometry, name) VALUES %s"
+    store_id = 0
+    for row in features.itertuples():
+        point = row.geometry.centroid if not isinstance(row.geometry, Point) else row.geometry
 
-    for _, row in features.itertuples():
-        point = row["geometry"].centroid if not isinstance(row["geometry"], Point) else row["geometry"]
-
-        if row["shop"] in ["supermarket", "grocery", "greengrocer"]:
+        if row.shop in ["supermarket", "grocery", "greengrocer"]:
             polygon = Polygon([
                 (point.x + 50 * math.cos(math.radians(angle)), point.y + 50 * math.sin(math.radians(angle)))
                 for angle in range(0, 360, 60)
@@ -414,21 +425,12 @@ def process_food_stores(
             ])
 
         map_elements.append(polygon.buffer(20))
-        store_tuples_strPoly.append((str(row["shop"]), str(polygon), str(row["name"])))
-        store_tuples_Poly.append((str(row["shop"]), polygon, str(row["name"])))
-    try:
-        extras.execute_values(cursor, food_stores_query, store_tuples_strPoly)
-    except psycopg2.Error as e:
-        logging.error(f"Databse insertion failed: {e}")
-        connection.rollback()
-        raise
+        # New format: (simulation_instance, simulation_step, shop, geometry, name, store_id)
+        store_tuples_strPoly.append((None, 0, str(row.shop), str(polygon), str(row.name), store_id))
+        store_tuples_Poly.append((None, 0, str(row.shop), polygon, str(row.name), store_id))
+        store_id += 1
     
-    return (STRtree(map_elements),store_tuples_Poly) , 
-
-
-# Call the function and get the spatial index of mapped elements
-map_elements_index = process_food_stores(CENTER_POINT, DIST, map_elements, cursor)
-
+    return (STRtree(map_elements),store_tuples_Poly, store_tuples_strPoly)  
 
 def create_households_table(cursor: psycopg2.extensions.cursor) -> str:
     """
@@ -443,15 +445,20 @@ def create_households_table(cursor: psycopg2.extensions.cursor) -> str:
     create_households_query = '''
     CREATE TABLE households (
         id NUMERIC,
-        polygon TEXT,
+        simulation_instance_id UUID,
+        simulation_step INTEGER,
+        centroid_wkt TEXT,
         income NUMERIC,
         household_size NUMERIC,
         vehicles NUMERIC,
         number_of_workers NUMERIC,
-        walking_time TEXT,
-        biking_time TEXT,
-        transit_time TEXT,
-        driving_time TEXT
+        walking_time NUMERIC,
+        biking_time NUMERIC,
+        transit_time NUMERIC,
+        driving_time NUMERIC,
+        food_score NUMERIC,
+        stores_within_1_mile NUMERIC,
+        closest_store_miles NUMERIC
     );
     '''
 
@@ -461,7 +468,9 @@ def create_households_table(cursor: psycopg2.extensions.cursor) -> str:
     household_query = """
     INSERT INTO households 
     (id,
-     polygon,
+     simulation_instance_id,
+     simulation_step,
+     centroid_wkt,
      income,
      household_size,
      vehicles,
@@ -469,13 +478,16 @@ def create_households_table(cursor: psycopg2.extensions.cursor) -> str:
      walking_time,
      biking_time,
      transit_time,
-     driving_time) 
+     driving_time,
+     food_score,
+     stores_within_1_mile,
+     closest_store_miles) 
     VALUES %s
     """
     return household_query
 
 
-def close_db_connection(connection: psycopg2.extensions.connection, cursor: psycopg2.extensions.cursor) -> None:
+def close_db_connection(connection: psycopg2.extensions.connection, cursor: psycopg2.extensions.cursor):
     """
     Commit any pending transactions and close the database connection and cursor.
 
@@ -558,26 +570,19 @@ def is_valid_house(
 ) -> bool:
     """
     Check whether a house polygon is valid by verifying it does not intersect
-    with roads or existing houses.
+    with roads, water bodies, or existing houses.
 
     Args:
         house (Polygon): The house polygon to check.
-        map_elements_index (STRtree): Spatial index for roads and store areas.
+        map_elements_index (STRtree): Spatial index for roads, stores, and water bodies.
         houses_index (rtree.index.Index): R-tree index of all placed houses.
-        map_elements (List[Polygon]): The list of buffered roads and stores.
+        map_elements (List[Polygon]): The list of buffered roads, stores, and water bodies.
 
     Returns:
         bool: True if house is valid; False if it intersects with other features.
     """
-    # Check collision with road/store polygons
-    search_area = Polygon([
-        (house.centroid.x + 10, house.centroid.y + 20),
-        (house.centroid.x - 10, house.centroid.y + 20),
-        (house.centroid.x - 10, house.centroid.y - 5),
-        (house.centroid.x + 10, house.centroid.y - 5)
-    ])
-
-    for index in map_elements_index.query(search_area):
+    # Check collision with road/store/water polygons using the actual house polygon
+    for index in map_elements_index.query(house):
         if map_elements[index].intersects(house):
             return False
 
@@ -624,7 +629,7 @@ def assign_household_attributes(
             "1 Person(s) 0 Worker(s)", "1 Person(s) 1 Worker(s)",
             "2 Person(s) 0 Worker(s)", "2 Person(s) 1 Worker(s)", "2 Person(s) 2 Worker(s)",
             "3 Person(s) 0 Worker(s)", "3 Person(s) 1 Worker(s)", "3 Person(s) 2 Worker(s)", "3 Person(s) 3 Worker(s)",
-            "4+ Person(s) 0 Worker(s)", "4+ Person(s) 1 Worker(s)", "4+ Person(s) 2 Worker(s)", "4+ Person(s) 3+ Worker(s)"
+            "4+ Person(s) 0 Worker(s)", "4+ Person(s) 1 Worker(s)", "4+ Person(s) 2 Worker(s)", "4+ Person(s) 3 Worker(s)"
         ]
     ]
 
@@ -748,7 +753,7 @@ def get_nearest_store(
 
     Args:
         house (Polygon): The house polygon to check
-        store_tuples: (List[Tuple[str, str, str]]): List of tuples (shop type, WKT polygon, name) for stores.
+        store_tuples: (List[Tuple[str, str, str]]): List of tuples (shop type, WKT polygon, name)
         shapely_loader (Any): Function to convert WKT string to Shapely geometry.
     
     Returns:
@@ -758,12 +763,19 @@ def get_nearest_store(
     store_distance = float('inf')
 
     for store in store_tuples:
-        store_poly = shapely_loader(store[1])
+        # Store format: (sim_instance, sim_step, shop, geometry, name, store_id)
+        # geometry is at index 3, could be Polygon or WKT string
+        geometry = store[3]
+        if isinstance(geometry, str):
+            store_poly = shapely_loader(geometry)
+        else:
+            store_poly = geometry  # Already a Polygon
+        
         dist = store_poly.distance(house)
 
         if dist <= store_distance:
             nearest_store = store_poly
-            store_distance = DIST
+            store_distance = dist  # Fixed: should use actual distance, not DIST
 
     return nearest_store
 
@@ -818,7 +830,7 @@ def process_housing_areas(
     Args:
         housing_areas (List[Polygon]): Polygons representing areas around residential roads.
         map_elements_index (STRtree): STRtree index for nearby road/store geometries.
-        map_elements (List[Polygon]): Buffered geometries (roads, stores) for intersection checking.
+        map_elements (List[Polygon]): Buffered geometries (roads, stores) for intersection checks.
         data (pd.DataFrame): Census tract data with attributes.
         store_tuples (List[Tuple[str, str, str]]): Store (shop type, WKT, name).
 
@@ -827,14 +839,19 @@ def process_housing_areas(
     """
     house_tuples: List[Tuple] = []
     total_count = 0
+    
+    # Debug counters
+    attempted = 0
+    failed_validation = 0
+    failed_tract = 0
+    failed_attributes = 0
+    failed_store = 0
+    success = 0
 
     # R-tree index to check house overlap
     houses_index = RTreeIndex()
     # STRtree index of tract geometries.
     tract_index = STRtree(data["geometry"])
-
-    vehicle_weights = List[int]
-    worker_weights = List[int]
     
     for i, housing_area in enumerate(housing_areas):
         logging.info(f"{round((i + 1) / len(housing_areas) * 100)}%")
@@ -855,27 +872,35 @@ def process_housing_areas(
                 )
 
                 house = create_house_polygon(location)
+                attempted += 1
 
                 if not is_valid_house(house, map_elements_index, houses_index, map_elements):
+                    failed_validation += 1
                     continue
 
                 houses_index.add(total_count, house.bounds)
 
                 tract = get_tract_for_house(house, tract_index, data)
                 if tract is None:
+                    failed_tract += 1
                     continue
 
                 try:
                     income, size, workers, vehicles = assign_household_attributes(
-                        tract, income_ranges, size_index_dict, workers_index_dict,
-                        vehicle_weights, worker_weights
+                        tract, income_ranges, size_index_dict, workers_index_dict
                     )
-                except Exception:
+                except Exception as e:
+                    failed_attributes += 1
+                    if failed_attributes == 1:  # Log first error
+                        logging.warning(f"First attribute assignment error: {e}")
                     continue
 
                 nearest_store = get_nearest_store(house, store_tuples, shapely.wkt.loads)
                 if nearest_store is None:
+                    failed_store += 1
                     continue
+                
+                success += 1
 
                 house_4326 = transform_polygon_coords(house, "EPSG:3857", "EPSG:4326")
                 store_4326 = transform_polygon_coords(nearest_store, "EPSG:3857", "EPSG:4326")
@@ -884,21 +909,42 @@ def process_housing_areas(
 
                 # Placeholder travel times (to be replaced with real data if available)
                 walking_time = biking_time = transit_time = driving_time = 0
+                
+                # Placeholder values for simulation-calculated fields
+                food_score = None
+                stores_within_1_mile = None
+                closest_store_miles = None
 
                 house_tuples.append((
                     total_count,
-                    str(house),
+                    None,  # simulation_instance_id (will be set during insertion)
+                    0,  # simulation_step (initial step)
+                    str(house.centroid),  # centroid_wkt instead of full polygon
                     income,
                     size,
                     vehicles,
                     workers,
-                    walking_time,
-                    biking_time,
-                    transit_time,
-                    driving_time
+                    str(walking_time),
+                    str(biking_time),
+                    str(transit_time),
+                    str(driving_time),
+                    food_score,
+                    stores_within_1_mile,
+                    closest_store_miles
                 ))
 
                 total_count += 1
+    
+    # Log summary statistics
+    logging.info("=" * 50)
+    logging.info("HOUSEHOLD GENERATION SUMMARY")
+    logging.info(f"Total houses attempted: {attempted}")
+    logging.info(f"Failed validation (intersects road/store/house): {failed_validation}")
+    logging.info(f"Failed tract check (outside census tract): {failed_tract}")
+    logging.info(f"Failed attributes (bad census data): {failed_attributes}")
+    logging.info(f"Failed store check (no nearby store): {failed_store}")
+    logging.info(f"Successfully created: {success}")
+    logging.info("=" * 50)
 
     return house_tuples
 
@@ -913,7 +959,7 @@ def insert_households(cursor, house_tuples: List[Tuple], household_query: str) -
         household_query (str): The SQL query template for inserting households
 
     Raises:
-        Exception: If insertion fails, the exception is re-raised after logging.infoing the error.
+        Exception: If insertion fails, the exception is re-raised after logging the error.
     """
     try:
         extras.execute_values(cursor, household_query, house_tuples)
@@ -951,41 +997,102 @@ def connect_to_db(
     )
     return conn, conn.cursor()
 
+#new function
+def initialize_simulation(
+    center_point: Tuple[float, float],
+    fips_code: str,
+    year: str,
+    dist: float,
+    api_key: str
+):
+    
+    state_code = fips_code[:2]
+    county_code = fips_code[2:]
+
+    county_data = fetch_county_data(households_key_list, str(year), state_code, county_code, api_key)
+    tract_data = load_and_merge_geodata(str(year), state_code, county_code, county_data)
+    map_elements, housing_areas, road_tuples = process_road_network(center_point, dist) 
+    store_index, store_tuples_poly, store_tuples_str = process_food_stores(center_point, dist, map_elements)
+    house_tuples = process_housing_areas(housing_areas, store_index, map_elements, tract_data, store_tuples_poly)
+
+    return {
+        'households' : house_tuples,
+        'stores' : store_tuples_str,
+        'roads' : road_tuples,
+        'tract_data' : tract_data,
+        'store_index' : store_index
+    }
+
 def main() -> None:
     """
     Main function to execute the full data processing and insertion pipeline.
     """
-    logging.info("Fetching household census data...")
-    county_data = fetch_county_data(households_key_list, YEAR, STATE_CODE, COUNTY_CODE, APIKEY)
-
-    logging.info("Loading and merging tract shapefile with census data...")
-    tract_data = load_and_merge_geodata(YEAR, STATE_CODE, COUNTY_CODE, county_data)
+    import uuid
+    
+    # new entry point
+    simulation_data = initialize_simulation(CENTER_POINT, FIPSCODE, YEAR, DIST, APIKEY)
 
     logging.info("Initializing database and creating tables...")
     connection, cursor = initialize_database_tables(HOST, NAME, USER, PASS, PORT)
     household_query = create_households_table(cursor)
+    
+    # Create or get the default simulation instance
+    logging.info("Creating simulation_instances table if not exists...")
+    cursor.execute('''
+        CREATE TABLE IF NOT EXISTS simulation_instances (
+            id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
+            name TEXT UNIQUE NOT NULL,
+            description TEXT,
+            created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
+        );
+    ''')
+    
+    # Insert or get default simulation instance
+    cursor.execute("""
+        INSERT INTO simulation_instances (name, description)
+        VALUES ('default_simulation', 'Brown County, Wisconsin')
+        ON CONFLICT (name) DO NOTHING
+        RETURNING id;
+    """)
+    result = cursor.fetchone()
+    if result:
+        simulation_instance_id = result[0]
+    else:
+        # If already exists, fetch it
+        cursor.execute("SELECT id FROM simulation_instances WHERE name = 'default_simulation';")
+        simulation_instance_id = cursor.fetchone()[0]
+    
+    logging.info(f"Using simulation_instance_id: {simulation_instance_id}")
 
-    logging.info("Processing road network...")
-    map_elements, housing_areas, road_tuples = process_road_network(CENTER_POINT, DIST)
+    logging.info("Inserting road data...")
+    roads_query = "INSERT INTO roads (name, highway, length, geometry, service) VALUES %s"
+    try:
+        extras.execute_values(cursor, roads_query, simulation_data['roads'])
+    except psycopg2.Error as e:
+        logging.error(f"Roads insertion failed: {e}")
+        connection.rollback()
+        raise
 
-    logging.info("Processing food stores...")
-    store_index = process_food_stores(CENTER_POINT, DIST, map_elements, cursor)
-
-    logging.info("Generating household records...")
-
-    # yeslai return wala banam, through process_food_stores()
-    store_tuples =  store_index[1] 
-
-    house_tuples = process_housing_areas(
-        housing_areas,
-        store_index,
-        map_elements,
-        tract_data,
-        store_tuples,
-    )
+    logging.info("Inserting food stores...")
+    food_stores_query = "INSERT INTO food_stores (simulation_instance, simulation_step, shop, geometry, name, store_id) VALUES %s"
+    # Update store tuples with simulation_instance_id
+    store_tuples_with_id = [(simulation_instance_id, step, shop, geom, name, sid) 
+                            for (_, step, shop, geom, name, sid) in simulation_data['stores']]
+    try:
+        extras.execute_values(cursor, food_stores_query, store_tuples_with_id)
+    except psycopg2.Error as e:
+        logging.error(f"Food stores insertion failed: {e}")
+        connection.rollback()
+        raise
 
     logging.info("Inserting households into the database...")
-    insert_households(cursor, house_tuples, household_query)
+    # Update household tuples with simulation_instance_id
+    household_tuples_with_id = [(hid, simulation_instance_id, step, centroid, income, size, vehicles, workers,
+                                 walking, biking, transit, driving, food_score, stores_1mi, closest_store)
+                                for (hid, _, step, centroid, income, size, vehicles, workers, 
+                                     walking, biking, transit, driving, food_score, stores_1mi, closest_store)
+                                in simulation_data['households']]
+    insert_households(cursor, household_tuples_with_id, household_query)    
 
     logging.info("Finalizing and closing connection...")
     connection.commit()
